+++ src/services/RGBLedService.cpp
@@ -1,5 +1,5 @@
 // RGBLedService.cpp
-#include "RGBLedService.h"
+#include "../include/services/RGBLedService.h"
 #include "config.h"
 
 #define CHANNEL_RED    0
