+++ src/ui/SunMoonUI.cpp
@@ -1,4 +1,4 @@
-#include "SunMoonUI.h"
+#include "../include/ui/SunMoonUI.h"
 #include "icons.h"
 #include "config.h"
